Prevent next/prev buttons being blocked by short sections
In particular, it was the last few "Community" sections. When using
next/prev buttons or scrolling normally, the 250px was too large and
caused the 'scroll spy' to stay on the previous section. If a user was
navigating this way, they'd be "blocked" from processing.

Reducing the offset pixels fixes the problem.